pyproject.toml
[project]
name = "aws-strands"
version = "0.2.1"
description = "Add your description here"
authors = [{ name = "James Dickson" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.0.1",
    "strands-agents-builder>=0.1.7",
    "strands-agents-tools>=0.2.1",
    "uipath>=2.0.82",
    "python-dotenv>=1.0.0",
]
